basic filtrelemeler yapıldı
